refactor: 발견 코드 위치 변경
- 컨트롤러와 서비스 Achievement로 통합
- DTO, Assembler는 각 도메인에 해당되는 파일 위치로 이동
