+++ src/main/java/site/beilsang/beilsang_server_v2/domain/achivement/service/AchievementService.java
@@ -1,0 +1,22 @@
+package site.beilsang.beilsang_server_v2.domain.achivement.service;
+
+import site.beilsang.beilsang_server_v2.domain.challenge.dto.res.HallOfFameListResDto;
+import site.beilsang.beilsang_server_v2.domain.feed.dto.res.PreviewFeedListResDTO;
+import site.beilsang.beilsang_server_v2.global.enums.Category;
+
+public interface AchievementService {
+
+    /**
+     * 특정 카테고리에 대한 명예의 전당을 조회합니다.
+     * @param category 조회할 카테고리
+     * @return 해당 카테고리의 명예의 전당 목록 (1위~10위)
+     */
+    HallOfFameListResDto getCategoryHallOfFame(Category category);
+    /**
+     * 특정 카테고리에 대한 피드들을 조회합니다.
+     * @param category 조회할 카테고리
+     * @param page 페이지 번호
+     * @return 해당 카테고리와 페이지에 해당하는 챌린지 피드 목록
+     */
+    PreviewFeedListResDTO getFeedsByCategory(Category category, Integer page);
+}
