--- src/main/java/site/beilsang/beilsang_server_v2/domain/feed/service/FeedService.java
@@ -1,9 +1,0 @@
-package site.beilsang.beilsang_server_v2.domain.feed.service;
-
-
-import site.beilsang.beilsang_server_v2.domain.feed.dto.res.PreviewFeedListResDTO;
-import site.beilsang.beilsang_server_v2.global.enums.Category;
-
-public interface FeedService {
-    PreviewFeedListResDTO getFeedsByCategory(Category category, Integer page);
-}
